rootfs-builder.jpl: fix stage names and don't store artifacts
Fix the name of the pipeline stages to not have unique names in every
job, and split the job in to "build" and "upload" stages.

Also don't archive artifacts in Jenkins, this takes space and the
files are not used since they are pushed directly to the KernelCI
storage server.

Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>