kci_rootfs: add debian riscv64 support for sid / debian-ports
riscv64 support is available in debian, but only in unstable/sid.  It
also requires using debian-ports.

In order to use debian-ports, the debian mirror, keyring package and
keyring file need to all be parameterized.

While adding support for sid, add a sid-specific list of removed
packages.  Copied from buster, but libunistring removed since it's
used in the minimal rootfs.

Signed-off-by: Kevin Hilman <khilman@baylibre.com>