test-configs: Enable clang boot testing
Currently we only build test with clang, we don't boot the resulting
binaries as there were some issue with the boot-specific reporting.  As
that is being phased out there is now no reason not to include clang
binaries in testing so do so, existing use of clang by developers and
results on staging do not suggest this is likely to result in many
additional failures.

Signed-off-by: Mark Brown <broonie@kernel.org>